feat: Add locale-specific layout component and files
