+++ apps/web/src/app/[locale]/page.tsx
